split up api.dart into serveral parts
