+++ pkgs/dart_mcp/lib/server.dart
@@ -2,5 +2,5 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-export 'src/api.dart';
+export 'src/api/api.dart';
 export 'src/server/server.dart';
